requirements.txt
Flask>=3.1.3
pymongo>=4.16.0
numpy>=2.4.4
scipy>=1.17.1
scikit-learn>=1.7.2
python-dotenv>=1.2.2
